moved logic to shared-module and added documentation
