+++ ports/nrf/mpconfigport.mk
@@ -52,8 +52,6 @@ MCU_SUB_VARIANT = nrf52840
 # Fits on nrf52840 but space is tight on nrf52833.
 CIRCUITPY_AESIO ?= 1
 
-CIRCUITPY_MSGPACK ?= 1
-
 SD ?= s140
 SOFTDEV_VERSION ?= 6.1.0
